helloflask/requirements.txt
flask
yfinance
gunicorn
python-dotenv
os
